test(file_type_download_support) : fix broken tests for FilePreloaderExecutors SDK-3913
